agent  login, aget dashboard , role system, authenticate route

+++ src/Components/Header.tsx
@@ -7,7 +7,9 @@ import SearchIcon from '@material-ui/icons/Search';
 import Typography from '@material-ui/core/Typography';
 import ExitToAppIcon from '@material-ui/icons/ExitToApp';
 import { useHistory } from "react-router-dom";
-
+import { BASE_URL } from "../Common/constant";
+import axios from "axios";
+import { ToastContext } from "../Common/ToastProvider";
 
 const useStyles = makeStyles((theme) => ({
   toolbar: {
@@ -30,6 +32,7 @@ const Header=()=> {
   const classes = useStyles();
   let history = useHistory();
   const [today, setDate] = React.useState(new Date());
+  const { showToast } = React.useContext(ToastContext);
 
   const hour = today.getHours();
   const wish = `Good ${(hour < 12 && 'Morning') || (hour < 17 && 'Afternoon') || 'Evening'}, `;
@@ -43,9 +46,34 @@ const Header=()=> {
   }
 
 
-  const handleLogout = () => {
-    localStorage.clear();
-    history.push("/");
+
+  const getToken = () => {
+    //@ts-ignore
+    return localStorage.getItem("access_token")
+}
+
+  const handleLogout = async () => {
+    try {
+        const result = await axios.post(BASE_URL + "user/logout",
+            {},
+            {
+                headers: {
+                    encryption: false,
+                    access_token: getToken()
+                }
+            })
+        if (result.data && result.data != null) {
+            showToast(result.data.message, "success");
+            localStorage.clear();
+            history.push("/");
+        }
+        else {
+            showToast(result.data.message, "error");
+        }
+    } catch (error) {
+        console.log(error)
+        showToast("Something went wrong!", "error");
+    }
 };
